pyproject.toml
[project]
name = "repl-nix-homeinventorytracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "wtforms>=3.2.1",
    "flask-session",
    "gunicorn",
    "python-dotenv",
]
